Added file processing to GenerateReport.py.
Takes in the tracked_numbers.txt file and reads in the tallied numbers
for max and 649. If missing returns an array of zeros.

Takes in the two .csv files and counts the number of lines. If the .csv
file is missing returns 0.

Generates the average times a number is picked using the tally and number
of times numbers have been drawn.