Finished up the report functions program. Moved some functions around.
GenerateReport.py: Will no longer count the number of lines in the csv.
It is expected that the tracked_numbers.txt file will contain the
necessary data. readTotal function expects the following format:
- line count for Lotto Max
- 7 lines of comma separated numbers, each line containing 50 numbers
- line count for Lotto 649
- 6 lines of comma separated numbers, each line containing 49 numbers
The program will now generate a separate bar graph for each number
position, showing the likelihood a number will be placed in that
position. SuggestNumbers function takes in the line count, the tally for
each position, possible choices and number guesses. The function will
use the tallies, line count and possible choices to generate a random
series of numbers that are weighted. The makeHTML function will use the
suggested numbers and images in the report.

GetLottoNumbers.py: Will now count the number of lines in the csv files.
Will now format the tracked_numbers.txt file in the proper format. Will
properly increment the tally in the txt file.

report.css: Formats the generated html file from GenerateReport.py.

.gitignore: Ignores any images generated from GenerateReport.py

GetLottoNumbers.py: